fix: add profiles for dist-all.
Because dist-all will contain all platforms libraries, it will conflict
with the same classes, MKL.class, Loader.class and BigQuant.class.

For maven, it will pass by default. But for sbt, it will need
mergeStragy to handle the conflicts. So we make the packaging as a
jar.

After it becomes a jar, the default `mvn clean package` will package it
on all the platforms. So we need an ugly command on each platform:

`mvn clean package -P linux -pl '!dist/dist-all'`

So add the profiles to handle this. So `mvn clean package -P linux`
will generate a `all-0.4.0-SNAPSHOT.jar` which only contains the linux
libraries.

`mvn clean package -P deploy` will generate a jar which contains all
libraries. But you should copy all pre-built libraries to
`target/classes` folder.